Flag & delete positionering
+ delete fine-tooning

+++ post.php
@@ -20,12 +20,12 @@
     $userData = $user->getUserDetailsByUserID($getUserID);
 
     // ALBERT JE MOET EEN SUBMIT KNOP KIEZEN :o
-	/*if(!empty($_POST['btnDeletePost'])){
+	if(!empty($_POST['btnDeletePost'])){
 		$deletePostID = $_POST['deletePostID'];
 		$post = new Post();
 		$post->deletePost($deletePostID);
 		header('Location: profile.php?profile='.$_SESSION['username'].'');
-	}*/
+	}
 
     // REDIRECT ON UNEXISTING POST used to prevent messing with url parameters on single post page
     if($post->checkComboExist($getPost, $getUserID) == false){
@@ -138,24 +138,46 @@
                 <input type="hidden" id="inputPostID" value="<?php echo $postData['id'];?>">
             </form>
         </div>
+        <div class="scheidingStreep"></div>
+        <div class="flagDeleteButton">
+        	<a class="btn-open" href="#">...</a>
+        </div>
+</div>
+<div class="flagDeleteWrap">
+      
+       <section class="flagDeleteMenu">
         <div class="commentsFlag--Individual">
             <form action="" method="POST">
                 <input type="hidden" name="postID" class="flagID" value="<?php echo $postData['id']; ?>">
                 <?php if($post->checkIfFlagged($postData['id']) == true): ?>
-                    <button type="submit" class="post__flag" name="flagPost"><span class="glyphicon <?php echo "f" . $postData['id']; ?> flagged glyphicon-flag"></span></button>
+                    <button type="submit" class="post__flag <?php echo "btn" . $postData['id']; ?> flagged" name="flagPost"><span class="glyphicon <?php echo "f" . $postData['id']; ?>"></span>Report inappropriate</button>
                 <?php else: ?>
-                    <button type="submit" class="post__flag" name="flagPost"><span class="glyphicon <?php echo "f" . $postData['id']; ?> glyphicon-flag"></span></button>
+                    <button type="submit" class="post__flag <?php echo "f" . $postData['id']; ?>" name="flagPost"><span class="glyphicon <?php echo "f" . $postData['id']; ?> "></span>Report inappropriate</button>
                 <?php endif; ?>
             </form>
         </div>
         <?php if($_SESSION['username'] == $userData['username']): ?>
-            <form action="" method="POST">
-                <input type="hidden" name="deletePostID" class="deleteID" value="<?php echo $postData['id']; ?>">
-                <button type="submit" class="post__delete" name="btnDeletePost" id="btnDeletePost"><span class="glyphicon glyphicon-trash"></span></button>
+        <div class="deleteOwnPost">
+               <form action="" method="POST">
+                <input type="hidden" name="deletePostID" id="deletePostID" class="deleteID" value="<?php echo $postData['id']; ?>">
+                <input type="submit" value="Delete this post" class="post__delete" name="btnDeletePost" id="btnDeletePost">
             </form>
+			</div>
         <?php endif; ?>
-    </div>
-
-</div>
+        <div class="cancelFlagDelete">
+        	<a href="#">Cancel</a>
+        </div>
+        </section>
+        </div>
+		</div>
+		<script type="text/javascript">
+			$('html').on("click", function(e){
+				if($(e.target).is('.flagDeleteButton a')){
+					$('.flagDeleteWrap').show();
+				}else if($(e.target) != $('.flagDeleteMenu')){
+					$('.flagDeleteWrap').hide();
+				}
+			});
+		</script>
 </body>
 </html>
